Added the implementation for cart layer

+++ ecommerce/src/main/java/com/example/ecommerce/model/CartItem.java
@@ -26,7 +26,7 @@ public class CartItem {
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 	
-	private int quantity;
+	private Long quantity;
 	
 	private BigDecimal price;
 	
@@ -37,7 +37,6 @@ public class CartItem {
 	@ManyToOne
 	@JoinColumn(name = "cart_id")
 	private Cart cart;
-	
 	
 
 }
